requirements.txt
absl-py==2.0.0
accelerate==0.26.1
aiodns==3.1.1
aiohttp==3.8.6
aiosignal==1.3.1
albumentations==1.3.1
annotated-types==0.6.0
anyio==4.1.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.7.2
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.1.0
Augmentor==0.2.12
Babel==2.14.0
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.1.0
blinker==1.7.0
Brotli==1.1.0
cachetools==5.3.2
catalyst==22.4
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
click==8.1.7
cloudscraper==1.2.71
colorama==0.4.6
comm==0.2.0
comtypes==1.2.0
contourpy==1.2.0
cryptography==41.0.5
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
Django==5.0.2
dm-tree==0.1.8
docopt==0.6.2
efficientnet_pytorch==0.7.1
et-xmlfile==1.1.0
executing==2.0.1
fastapi==0.110.0
fastjsonschema==2.19.0
fernet==1.0.1
filelock==3.13.1
fileupload==0.1.5
Flask==3.0.2
Flask-Cors==4.0.0
flaskwebgui==1.1.1
flatbuffers==24.3.25
fonttools==4.47.2
fqdn==1.5.1
frozenlist==1.4.0
fsspec==2023.12.2
gast==0.5.4
google-auth==2.26.2
google-auth-oauthlib==1.2.0
google-pasta==0.2.0
grpcio==1.60.0
gunicorn==21.2.0
h11==0.14.0
h5py==3.12.1
huggingface-hub==0.20.2
hydra-slayer==0.5.0
idna==3.4
imageio==2.33.1
install==1.3.5
ipykernel==6.27.1
ipython==8.12.3
ipywidgets==8.1.1
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.9.0
jupyter-lsp==2.2.1
jupyter_client==8.6.0
jupyter_core==5.5.0
jupyter_server==2.11.1
jupyter_server_terminals==0.4.4
jupyterlab==4.0.9
jupyterlab-widgets==3.0.9
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.2
keras==3.6.0
kiwisolver==1.4.5
lazy_loader==0.3
libclang==16.0.6
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.3
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.2
ml-dtypes==0.4.1
mpmath==1.3.0
multidict==6.0.4
namex==0.0.7
nbclient==0.9.0
nbconvert==7.11.0
nbformat==5.9.2
nest-asyncio==1.5.8
networkx==3.2.1
nltk==3.8.1
notebook==7.0.7
notebook_shim==0.2.3
numpy==1.26.4
oauthlib==3.2.2
opencv-python==4.9.0.80
opencv-python-headless==4.9.0.80
openpyxl==3.1.2
opt-einsum==3.3.0
optree==0.13.0
outcome==1.2.0
overrides==7.4.0
packaging==23.2
pandas==2.1.4
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
pillow==10.2.0
pipreqs==0.5.0
platformdirs==4.0.0
price-parser==0.3.4
prometheus-client==0.19.0
prompt-toolkit==3.0.41
protobuf==4.23.4
psutil==5.9.6
pure-eval==0.2.2
pyaes==1.6.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
PyAudio==0.2.14
pycares==4.4.0
pycparser==2.21
pydantic==2.6.4
pydantic_core==2.16.3
Pygments==2.17.2
pyparsing==3.1.1
pypiwin32==223
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-json-logger==2.0.7
pyttsx3==2.90
pytz==2023.3.post1
pywin32==306
pywinpty==2.0.12
PyYAML==6.0.1
pyzmq==25.1.1
qtconsole==5.5.1
QtPy==2.4.1
qudida==0.0.4
referencing==0.31.1
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.0
rpds-py==0.13.2
rsa==4.9
ruamel.yaml==0.17.35
ruamel.yaml.clib==0.2.8
safetensors==0.4.1
scikit-image==0.22.0
scikit-learn==1.3.2
scipy==1.11.4
seaborn==0.13.2
selenium==4.14.0
Send2Trash==1.8.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
SpeechRecognition==3.10.0
sqlparse==0.4.4
stack-data==0.6.3
starlette==0.36.3
sympy==1.12
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
tensorflow==2.18.0
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.31.0
tensorflow_intel==2.18.0
termcolor==2.4.0
terminado==0.18.0
threadpoolctl==3.2.0
tifffile==2023.12.9
tinycss2==1.2.1
torch==2.2.1
torch-tb-profiler==0.4.3
torchvision==0.17.1
torchxrayvision==1.2.1
tornado==6.4
tqdm==4.66.1
traitlets==5.14.0
trio==0.22.2
trio-websocket==0.11.1
types-python-dateutil==2.8.19.14
typing_extensions==4.9.0
tzdata==2023.3
updates==0.1.7.1
uri-template==1.3.0
urllib3==2.0.6
uvicorn==0.29.0
waitress==3.0.0
wcwidth==0.2.12
webcolors==1.13
webdriver-manager==4.0.1
webencodings==0.5.1
websocket-client==1.6.4
Werkzeug==3.0.1
widgetsnbextension==4.0.9
wikipedia==1.4.0
wrapt==1.14.1
wsproto==1.2.0
yarg==0.1.9
yarl==1.9.2
